+++ sanitize_html/CHANGELOG.md
@@ -1,3 +1,7 @@
+## v1.1.0
+ * Add options `allowElementId` and `allowClassName` to allow specific element
+   ids and class names.
+ 
 ## v1.0.0
  * Initial release.